u8 buffer working

+++ src/lib/ppu.pipleline.c
@@ -8,7 +8,7 @@ bool window_visible() {
         lcd_get_context()->win_y >= 0 && lcd_get_context()->win_y < YRES;
 }
 
-void pixel_fifo_push(u32 value) {
+void pixel_fifo_push(u8 value) {
     fifo_entry *next = malloc(sizeof(fifo_entry));
     next->next = NULL;
     next->value = value;
@@ -24,7 +24,7 @@ void pixel_fifo_push(u32 value) {
     ppu_get_context()->pfc.pixel_fifo.size++;
 }
 
-u32 pixel_fifo_pop() {
+u8 pixel_fifo_pop() {
     if (ppu_get_context()->pfc.pixel_fifo.size <= 0) {
         fprintf(stderr, "ERR IN PIXEL FIFO!\n");
         exit(-8);
@@ -34,13 +34,13 @@ u32 pixel_fifo_pop() {
     ppu_get_context()->pfc.pixel_fifo.head = popped->next;
     ppu_get_context()->pfc.pixel_fifo.size--;
 
-    u32 val = popped->value;
+    u8 val = popped->value;
     free(popped);
 
     return val;
 }
 
-u32 fetch_sprite_pixels(int bit, u32 color, u8 bg_color) {
+u8 fetch_sprite_pixels(int bit, u8 color, u8 bg_color) {
     for (int i=0; i < ppu_get_context()->fetched_entry_count; i++) {
         int sp_x = (ppu_get_context()->fetched_entries[i].x - 8) + 
             ((lcd_get_context()->scroll_x % 8));
@@ -98,7 +98,7 @@ bool pipeline_fifo_add() {
         int bit = 7 - i;
         u8 hi = !!(ppu_get_context()->pfc.bgw_fetch_data[1] & (1 << bit));
         u8 lo = !!(ppu_get_context()->pfc.bgw_fetch_data[2] & (1 << bit)) << 1;
-        u32 color = lcd_get_context()->bg_colors[hi | lo];
+        u8 color = lcd_get_context()->bg_colors[hi | lo];
 
         if (!LCDC_BGW_ENABLE) {
             color = lcd_get_context()->bg_colors[0];
@@ -250,7 +250,7 @@ void pipeline_fetch() {
 
 void pipeline_push_pixel() {
     if (ppu_get_context()->pfc.pixel_fifo.size > 8) {
-        u32 pixel_data = pixel_fifo_pop();
+        u8 pixel_data = pixel_fifo_pop();
 
         if (ppu_get_context()->pfc.line_x >= (lcd_get_context()->scroll_x % 8)) {
             ppu_get_context()->video_buffer[ppu_get_context()->pfc.pushed_x +
